fix: move calls list table header into component to simplify calls list template
